requirements.txt
eventlet==0.33.2
pygame==2.1.2
python-socketio==5.7.2
requests==2.28.1
pygame-menu==4.2.8